% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispfit-package.R
\docType{package}
\name{dispfit-package}
\alias{dispfit}
\alias{dispfit-package}
\title{dispfit: A package to Fit Distributions to Dispersal Data}
\description{
The dispfit package fits several pre-defined distributions to dispersal or movement data,
computing several estimators: AIC, AICc, BIC, Chi-squared, and Kolgomorov-Smirnov.
It also estimates the parameter(s) value(s) and SE of each distribution,
as well as its Mean, Variance, Skewness, and Kurtosis.
}
\details{
The dispfit package fits 9 well-known distributions for estimating dispersal kernels
(Clark et al., 1999; Nathan et al., 2012) (Table 1).
The simplest functions considered are the single‐parameter Rayleigh and Exponential,
which are particularly popular in mathematical developments of theory concerning spatial dynamics
(O'Dwyer & Green 2010; Gilbert et al. 2014; Harsch et al. 2014).
The remaining 7 functions are two-parameter distributions which are often referred to better
represent real dispersal kernels than Rayleigh and Exponential functions
(Bullock and Clarke, 2000; Clark et al., 1999).
}
\section{Package functions}{

dispfit includes two main functions:
\tabular{ll}{
  \code{\link{dispersal.kernel}} \tab Fits several pre-defined distributions to dispersal or movement data. \cr
  \code{\link{plot.dispfit}} \tab Plots the distributions previously fitted by dispersal.kernel. \cr
}
}

\section{Probability density function}{

Assuming that a single point is the origin site of all dispersers,
then the dispersal distance of each disperser is the Euclidian distance between the origin and its end point.
The dispersal distances of all dispersers reflect a continuous parametric distribution,
or probability density function (pdf), that characterizes the studied population.
A dispersal kernel is then defined as the pdf of the distribution of the values
of the Euclidean distances between the source and the final location of a dispersal event.
There are several characterizations of a dispersal kernel,
for instance Nathan et al. (2012) distinguish between “dispersal distance kernel, KD”,
and “dispersal location kernel, KL”.
}

