% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersal.kernel.plot.R
\name{plot.dispfit}
\alias{plot.dispfit}
\title{Plots from dispersal kernel fit for dispersal data}
\usage{
\method{plot}{dispfit}(
  data,
  fit.criteria = NULL,
  criteria.dif = 2,
  envelopes = TRUE,
  plot.data = TRUE
)
}
\arguments{
\item{data}{Output object from the \code{\link{dispersal.kernel}} function.}

\item{fit.criteria}{Either a numeric value referring to the number of distributions to plot (ordered according to the order.by argument from dispersal.kernel); a character string choosing to plot “all” distributions, to plot the distributions with an “AIC”, “AICc”, or “BIC” difference to the top distribution bellow the number set by criteria.dif, or to plot specific distributions. By default, only the top distribution is plotted.}

\item{criteria.dif}{A numeric value used when “AIC”, “AICc”, or "BIC" are also selected in fit.criteria. Its value refers to the estimator difference to the top distribution. Only distributions with their estimator difference to the top model falling below this value will be plotted. The default is 2.}

\item{envelopes}{A boolean (TRUE/FALSE) value choosing whether to plot the confidence envelopes associated with each distribution. Default is TRUE.}

\item{plot.data}{A boolean (TRUE/FALSE) value selecting to plot the density of the data.}
}
\description{
Plots the distributions previously fitted by \code{\link{dispersal.kernel}} against the data density plot.
}
\examples{

sim <- rexp(100, rate = 0.01)

test <- dispersal.kernel(sim)

plot(test)
}
\seealso{
\code{\link{dispersal.kernel}}
}
\keyword{kernel}
\keyword{plot}
